+++ be/src/main/java/teamproject/airbnb/controller/RoomController.java
@@ -1,4 +1,4 @@
-package teamproject.airbnb.api;
+package teamproject.airbnb.controller;
 
 import java.time.LocalDate;
 import lombok.RequiredArgsConstructor;
@@ -8,8 +8,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
-import teamproject.airbnb.api.dto.RoomQuantityResponseDto;
-import teamproject.airbnb.api.dto.RoomSimpleInfoResponseDto;
 import teamproject.airbnb.service.RoomService;
 
 @RestController
