+++ be/src/main/java/teamproject/airbnb/service/RoomQuantityDto.java
@@ -1,8 +1,7 @@
 package teamproject.airbnb.service;
 
-import java.util.Comparator;
+import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import teamproject.airbnb.domain.Room;
@@ -13,25 +12,61 @@ public class RoomQuantityDto {
 
 	private Long minimumPrice;
 	private Long maximumPrice;
-	private Long[] roomQuantity;
+	private List<Long> priceRange;
+	private List<Long> roomQuantity;
 
 	public static RoomQuantityDto from(List<Room> rooms) {
 
-		Long maximumPrice = rooms.stream()
+		// 가격을 기준으로 내림차순으로 정렬한 배열
+		Long[] priceArray = rooms.stream()
 			.mapToLong(x -> x.getPrice())
-			.max()
-			.getAsLong();
-
-		Long minimumPrice = rooms.stream()
-			.mapToLong(x -> x.getPrice())
-			.min()
-			.getAsLong();
-
-		// TODO 동일한 값은 같은 인덱스에 넣어주도록 해야 한다.
-		Long[] roomQuantity = rooms.stream()
-			.sorted(Comparator.comparing(Room::getPrice))
+			.sorted()
+			.boxed()
 			.toArray(Long[]::new);
 
-		return new RoomQuantityDto(minimumPrice, maximumPrice, roomQuantity);
+		Long maximumPrice = priceArray[0];
+		Long minimumPrice = priceArray[priceArray.length - 1];
+
+		// 가격 중에 중복되는거 제거 해서 가격범위 만들기
+		Long[] priceRange = rooms.stream()
+			.mapToLong(x -> x.getPrice())
+			.sorted()
+			.distinct()
+			.boxed()
+			.toArray(Long[]::new);
+
+		//===================
+
+		Long[] roomQuantity = new Long[priceRange.length];
+		int count = 0;
+		for (int i = 0; i < priceArray.length; i++) {
+
+			// 방개수를 ++
+			roomQuantity[count]++;
+			if (i + 1 < priceArray.length) {
+				// 현 인덱스와 다음 인덱스의 값이 다르면 count를 ++
+				if (priceArray[i] != priceArray[i + 1]) {
+					count++;
+				}
+			}
+		}
+
+		return new RoomQuantityDto(minimumPrice, maximumPrice, Arrays.stream(priceRange).toList(),
+			Arrays.stream(roomQuantity).toList());
+	}
+
+	public String getMinimumPriceFormat() {
+		return this.getMinimumPrice().toString()
+			.replaceAll("\\B(?=(\\d{3})+(?!\\d))", ",");
+	}
+
+	public String getMaximumPriceFormat() {
+		return this.getMaximumPrice().toString()
+			.replaceAll("\\B(?=(\\d{3})+(?!\\d))", ",");
+	}
+
+	public String getAveragePriceFormat() {
+		Long averagePrice = (this.getMinimumPrice() + this.getMaximumPrice()) / 2;
+		return averagePrice.toString().replaceAll("\\B(?=(\\d{3})+(?!\\d))", ",");
 	}
 }
